requirements.txt
pyjwt
fastapi
httpx
pandas
pg8000
pydantic
requests
SQLAlchemy
uvicorn[standard]